% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_delayed.R
\name{count_delayed}
\alias{count_delayed}
\title{Title Count delayed flights}
\usage{
count_delayed(data)
}
\arguments{
\item{data}{dataframe to work at}
}
\value{
count_delays number of flights that were delayed
}
\description{
Title Count delayed flights
}
\examples{
count_delayed(read_data())
}
