% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Title Read data file (csv) from directory by the user}
\usage{
read_data()
}
\value{
dataframe
}
\description{
Title Read data file (csv) from directory by the user
}
\examples{
read_data()
read_data("C:\\\\Users\\\\Pushpinder\\\\BTIIC\\\\Code\\\\Data","git-Customer-Churn.csv")
}
